Creates encrypted cipher stream
This commit includes encryption to messages. Initial messages are
allowed to start un-encrypted, when written. When a new message is read,
the information received can be used to initialize the stream.

If there is no encryption happening, it will not encrypt the buffer
produce unencrypted.

This used a share-able Cipher struct, allowing for the read and writer
control when it should be encrypted. This means that the initiator can
send an unencrypted message, and wait for the encrypted message using
the information sent.

An example is provided, showing it can be run both by an uncripted
client and a encrypted client.

Example:

```sh
cargo run --example tcp_encrypted -- server localhost:8080
``

```sh
cargo run --example tcp_encrypted -- server localhost:8080
cargo run --example tcp -- client localhost:8080
```

Only XSalsa20 is supported at this moment. A trait could be provided to
extend the design to other ciphers, but this is not proposed here.

Related to: https://github.com/datrs/simple-message-channels/issues/3