Remove internal buffering
Using the split strategy to upgrade the underlying `Reader<R>` from an
plain channel to an encrypted channel caused some data-loss.

The plain channel inside the `Reader` would also buffer encrypted
bytes that would go missing from the encrypted `Reader`.

From the docs, we can see this is an issue:
```
When the BufReader<R> is dropped, the contents of its buffer will be
discarded. Creating multiple instances of a BufReader<R> on the same
stream can cause data loss.
```
https://doc.rust-lang.org/std/io/struct.BufReader.html

This commit removes the `BufReader` and `BufWriter` from the crate. This
moves the responsability of the caller to wrap the encrypted channel
into a BufReader/BufWriter and leave the plain channel unbuffered.